HW3 Invalid login test
