--- SavchenkoProj/src/test/java/org/loginTest/LoginTestWithPageObject.java
@@ -1,16 +1,0 @@
-package org.loginTest;
-
-import org.baseTest.BaseTest;
-import org.junit.Test;
-
-public class LoginTestWithPageObject extends BaseTest {
-
-    @Test
-    public void T0001_validLogin() {
-        pageProvider.getLoginPage().openPage();
-        pageProvider.getLoginPage().enterTextIntoInputLogin("qaauto");
-        pageProvider.getLoginPage().enterTextIntoInputPassword("123456qwerty");
-        pageProvider.getLoginPage().clickOnButtonSignIn();
-        pageProvider.getHomePage().checkIsButtonSignOutVisible();
-    }
-}
